feat(cart): add cart features and update global exception handling

+++ demo/src/main/java/com/project/demo/repository/ProductRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import com.project.demo.model.Product;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer>, JpaSpecificationExecutor<Product> {
+public interface ProductRepository extends JpaRepository<Product, Long>, JpaSpecificationExecutor<Product> {
 
 	Optional<Product> findByUuid(UUID uuid);
 }
